chore: move TabHistory from widgets to util
